+++ Frontend/src/Components/Cat-itemall/SteelAll/steelcategory.jsx
@@ -8,6 +8,7 @@ import config from "../../../config";
 import Navbar from "../../Navbar/Navbar";
 import Footer from "../../Footer/Footer";
 import './steelall.css'; 
+import { Snackbar } from '@mui/material';
 
 const CategorySteelall = () => {
   const [data, setData] = useState([]);
@@ -17,7 +18,7 @@ const CategorySteelall = () => {
   const [likeCounts, setLikeCounts] = useState({});
   const [searchQuery, setSearchQuery] = useState(""); // State for search query
   const navigate = useNavigate();
-
+  const [snackbarOpen, setSnackbarOpen] = useState(false);
   const steelRoute = `${config.apiURL}/steelRoute/steel`;
 
   useEffect(() => {
@@ -60,6 +61,10 @@ const CategorySteelall = () => {
     return localStorage.getItem('userId');
   };
 
+  const handleSnackbarClose = () => {
+    setSnackbarOpen(false);
+  };
+
   const handleCardClick = (steelId) => {
     navigate(`/steelview/${steelId}`);
   };
@@ -68,6 +73,14 @@ const CategorySteelall = () => {
     const userId = getUserId();
     const productId = steelId;
 
+    if (!userId) {
+
+      setSnackbarOpen(true);
+      setTimeout(() => {
+        navigate('/Login'); 
+      }, 1500);
+      return; 
+    }
     try {
       if (favourites.includes(productId)) {
         await axios.delete(`${config.apiURL}/favourites/remove`, {
@@ -114,15 +127,15 @@ const CategorySteelall = () => {
           <h2>Steel Products</h2>
         </div>
 
-        <div className="cat-search-container">
+        <div className="steeleall-search-container">
           <input
             type="text"
             placeholder="Search by steel type, brand, or category..."
             value={searchQuery}
             onChange={(e) => setSearchQuery(e.target.value)}
-            className="cat-search-input"
+            className="steeleall-search-input"
           />
-          <button className="cat-search-button">
+          <button className="steeleall-search-button">
             Search
           </button>
         </div>
@@ -191,6 +204,22 @@ const CategorySteelall = () => {
             })
           )}
         </div>
+        <Snackbar
+        open={snackbarOpen}
+        onClose={handleSnackbarClose}
+        message="You need to log in to add to favourites."
+        autoHideDuration={2000}
+        sx={{
+          '& .MuiSnackbarContent-root': {
+            backgroundColor: '#f44336', 
+            color: '#fff', 
+            borderRadius: '8px', 
+            padding: '11px',
+            fontSize: '0.8rem', 
+            boxShadow: '0 2px 10px rgba(0,0,0,0.2)', 
+          },
+        }}
+      />
       </div>
       <Footer />
     </>
